requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
google-generativeai
python-dotenv
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]